SHT-032 Move login to home
