+++ package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "clsx": "^2.1.1",
     "flowbite-react": "^0.9.0",
+    "jotai": "^2.8.0",
     "next": "14.2.2",
     "react": "^18",
     "react-dom": "^18",
