+++ src/app/atoms/index.tsx
@@ -1,0 +1,5 @@
+import { atom } from 'jotai';
+
+export const nameAtom = atom<string | null>(null);
+export const unitAtom = atom<string | null>(null);
+
